Create and Write or Load In the Filtered eBird Data

Filter the eBird data and sampling events using R package AUK.

munge_ebird_data(
  fns.ebird,
  dir.ebird.out,
  overwrite = FALSE,
  countries = NULL,
  states = NULL,
  complete.only = TRUE,
  protocol = c("Traveling", "Stationary"),
  species = "Double-crested Cormorant",
  remove.bbs.obs = TRUE,
  years = NULL,
  max.effort.km = NULL,
  max.effort.mins = NULL,
  max.num.observers = 10,
  f_obs_out = paste0(dir.ebird.out, "ebird_obs_filtered.txt"),
  f_samp_out = paste0(dir.ebird.out, "ebird_samp_filtered.txt")
)

Arguments

fns.ebird

File paths for the EBD and SamplingEvents data to import. Character vector of filenames for original files.

dir.ebird.out

Location of where to save and/or find the filtered/subsetted eBird data.

overwrite

logical If true will overwrite existing filtered data objects in directory 'dir.ebird.out'

countries

Character vector of country abbreviations for filtering ebird data against.

states

Character vector of state/province abbreviations for filtering ebird data against.

complete.only

logical if TRUE will remove incomplete checklists

protocol

which eBird protocol to include

species

species for filtering eBird data

remove.bbs.obs

logical if TRUE will use crude and untested method for removing BBS data from eBird data (see function for details)

years

years to include in data

max.effort.km

maximum distance (kilometers) of birding events to include

max.effort.mins

maximum number of minutes of birding events to include

max.num.observers

maximum number of observers declared at a birding event to include

f_obs_out

Filename for saving the filtered eBird observations data

f_samp_out

Filename for saving the filtered eBird sampling data