Partition the eBird Sampling Events Data File by States/Provinces

Relies heavily on package 'data.table' to import, filter, and write the sampling events data for each country of interest.

partition_ebird_events(
  dir.ebird.in,
  mmyyyy,
  outpath = NULL,
  overwrite = FALSE,
  out.filetype = ".csv.gz",
  countries = c("US", "CA", "MX"),
  ncores = NULL
)

Arguments

dir.ebird.in

directory path to where ebird sampling events is/are saved

mmyyyy

month and year associated with the sampling events file.

outpath

path location to where output files will be saved. If NULL will save to a directory within dir.ebird.in for each mmyyyy

overwrite

if TRUE, will overwrite existing partitioned file for country and mmyyyy combination

out.filetype

if writing new files, what file extension to search for? (e.g., .csv, .txt, .csv.gz).

countries

list of 2-letter country codes. Only sampling events from these countries will be partitioned into indiviudal sampling event files on local disk. Specify countries=NULL to partition all countries (this will take a while)

ncores

number of processors to engage during data import and export (using data.table)