Run BUGS Model in nimble

Run BUGS Model in nimble

Usage

run_in_nimble(
  myData,
  myModel,
  myInits = NULL,
  traceplots = TRUE,
  constants = NULL,
  monitors = NULL,
  savedir = "outputs",
  seed = sample(1:111111, size = 1),
  parallel = FALSE,
  verbose = TRUE,
  mod.name = "mynimbleModel",
  mcmc.specs = set_mcmc_specs(dev.mode = FALSE),
  ...
)

Arguments

myData

list List of elements comprising data (and constants if not provided separately) for use in nimble

myModel

nimbleCode object or filepath to a BUGS model.

myInits

list List of initial values.

traceplots

logical If TRUE will attempt to print default traceplots for model output

constants

optional Can be provided separately from myData. If both are provided, function will combine prior to inclusion in readNimbleModel

monitors

optional Character vector of parameters to monitor.

savedir

If not specified, will save the samples resulting from nimbleUI::nimble() to current working directory

seed

optional If specified will set a seed for random number generation.

parallel

logical If TRUE will run in parallel

verbose

logical Argument used in nimble::compileNimble(showCompilerOutput). If TRUE will generate more messages/information during sampling phase. Note: If parallel=TRUE, messages are suppressed given behavior of parallel compute.

mod.name

optional Used to save model output to file. Defaults to 'mynimbleModel'

mcmc.specs

list List of specs for use in MCMC specification (e.g., n.chains, n.adapt, n.cores). If any necessary values are missing, will use set_mcmc_specs to grab a default value.

...

Additional arguments